DockerFR/requirements.txt
fastapi
pydantic
uvicorn[standard]
python-multipart
opencv-python-headless
numpy
pillow
scikit-learn
insightface
onnxruntime